Add documentation for clarity of responsibilities
