+++ README.md
@@ -19,42 +19,36 @@ pip install pubsubs
 
 ### Publisher
 
-Configuring a Kafka publisher
-
-```yaml
-kafka:
-    backend: kafka
-    listeners::
-        - localhost:9092
-    poll: 0
-    message.timeout.ms: 1500
-```
-
 New Kafka publisher instance
 
 ```python
 from pubsubs.registry import Registry
 
-CONFIG = <see-above>
+config = {
+    "listeners": ['localhost:9092'],
+    "poll": 1.0,
+    "message.timeout.ms": 1500,
+}
 
 registry = Registry()
-registry.register_from_config(CONFIG)
 
-kafka = registry["kafka"]
+kafka = registry.new("kafka1", backend="kafka", **config)
+
 kafka.publish(topic="MyTopic", message="hey!")
 ```
 
 ### Subscriber
 
-Configuring a Kafka subscriber
+Configuring a Kafka subscriber and using a yaml.
 
 ```yaml
-kafka:
-    backend: kafka
-    listeners: ['localhost:9092']
-    poll: 0.1
-    group.id: 'myGroup'
-    auto.offset.reset: 'earliest'
+pubsubs:
+    kafka:
+        backend: kafka
+        listeners: ['localhost:9092']
+        poll: 0.1
+        group.id: 'myGroup'
+        auto.offset.reset: 'earliest'
 ```
 
 New Subscriber instance
@@ -65,6 +59,8 @@ from pubsubs.registry import Registry
 CONFIG = <see-above>
 
 registry = Registry()
+
+# This is alternative to using 'Registry.new'
 registry.register_from_config(CONFIG)
 
 kafka = registry["kafka"]
@@ -74,3 +70,13 @@ while True:
     message = subscriber.listen()
     print(message)
 ```
+
+### Configuration
+
+The subsection `pubsubs` in the yaml allows the config to be used along side
+[omniduct](https://github.com/airbnb/omniduct/blob/master/example_wrapper/example_wrapper/services.yml)
+
+| Name | Description | Example |
+| ---- | ----------- | ------- |
+| `backend` | Name of pub sub implementation | `backend: kafka` |
+| `listeners` | List of serving addresses | `listeners: ['localhost:9092']` |
