+++ pubsubs/__version__.py
@@ -17,3 +17,7 @@ __description__ = "Publish/Subscribe interface"
 
 __title__ = "pubsubs"
 __dependencies__ = ["interface_meta>=1.1.0"]
+__optional_dependencies__ = {
+    # Kafka client
+    "kafka": ["confluent-kafka>=1.0.0"]
+}
