+++ pubsubs/backends.py
@@ -1,2 +1,3 @@
+""" Registered concrete implementations of Pub/Sub."""
 # from .avro import AvroClient
 from .kafka import KafkaClient, KafkaSubscriber
